Rename generateTranslatedResource to translateResource in filters
The name more accurately describes the method, which starting from a source resource,
it generates a translated resource. This is also in preparation of a possible new method
(aptly called generateTranslatedResource maybe) where we could allow different
formats for source and target and thus starting from a source in the same format
wouldn't be possible.