Refactor json job store to use new structure
Previous version was error prone and not very scalable as it relied on a giant jobs.json that was used as a metadata index.
That made it hard to manually edit and remove jobs.

The new version gets rid of this file in favor of:
(1) a nested directory per language pair that helps querying all jobs for a specific pair. It also helps scaling to thousands of files.
(2) placing a status suffix (req/wip/done) to the json filename so that we can infer the job status looking at filenames without needing
to open and parse the entire files (if the done file is there, then job is done, if not and wip is there, then it's in flight, and if only req
is there it means the job was created but it failed).

The refactor is a step towards immutability as req and done are immutable. wip is still mutable because of the blocked status
and because of partial pulls.

The side-effects of the refactor:
1 - the numeric jobId is now gone and only the jobGuid is used
2 - as the jobId is gone, Visicode can't encode indexes anymore
3 - deprecated the sql stores as their usefulness is questionable at this stage
4 - previous translations jobs need to be refactored to the new scheme