Add option to export tm in job format
This can be used to dump the consolidated tm into a single job, thus reducing the number of files.

It also tries to clean up TU's to a fixed set of properties to balance off bugs that may have produced
wrong or incomplete properties.

TU's in flight are ignored. It's recommended to launch this command when there are no pending jobs.

Even though technically only the "done" job is necessary to populate the TM, a "req" job is also
generated to preserve notes that could be used in a bugfix project.