Add new TOSRefresh provider and bugfix option to push
This is the initial working version of a bugfix solution. It adds a new bugfixFilters config property to provide custom logic
to decide what to send to the specific bugfix provider.

The new TOSRefresh provider is a synchronous translator that checks the desired TU's against the latest in TOS.
If content_hash in TOS differs from the th property in the tu, the new entry is downloaded and a new job created.

The push command now has 2 new options: bugfix, bugfixmode. The former to select which bugfixFilters to use,
and the latter to select whether to scan all sources or all TM entries when evaulating what to send to bugfix.
Scanning the source may be helpful to bugfix only active entries or to recover a lost TM. Scanning the TM may
be useful to fix even outdated entries that are no longer in use.