Support pulling partially completed jobs
By default, incomplete jobs are discarded. The new --partial option allows to commit what
was complete and create a new pending job for the remaining units.

Also fixed a problem with TOS queries where the combination of id_order and id_content
is interpreted with an OR condition and not an AND. This means that now the latest translation
is always the one fetched and not the translation corresponding to the specific order.

Also, remove num property from jobs as it's now obsolete.