Revamp analyze command to make it extensible
Analyze command now takes an argument that specifies which analyzer to run and optional
extra custom parameters for the analyzer.
Analyzers can specify whether to analyze source or tm (used to be source only).
If it's TM-based, an option can limit to one language, otherwise all possible languages are looped through.
Analyzers can output structured or unstructured reports. If structured, they are in columnar format.
Some columns can be designated as groupable and are de-duped in the console output.
An --output option allows to write the report into a text/CSV file instead of the console.
4 built-in analyzers are provided: DuplicateSource, SmellySource, TextExpansionSummary, FindByExpansion.
Config allows to plug custom analyzers on top of the system-defined ones.