Refactor CLI to be scriptable
Move all reusable code from `l10n.js` to `src`. Now `l10n.js` should be the only file depending on
`commander` and defining CLI commands and options. It's also the only file that implements the
logic that looks for the l10nmonster.mjs config among ancestors of the current directory.

The new `cli.js` module contains all CLI commands that include console outputs, but doesn't
depend on commander and doesn't need to be invoked from the shell. This means that it can
be invoked from code by simply passing an instance of monster manager and an options object.

The new `defaultMonster.js` is a utility that can be used to instantiate a monster manager given
a config. It includes all default helpers like `l10n.js` used to do.

The regression script is now refactored to invoke all commands in a project from a node script
instead of the shell. This avoids multiple initializations for the same project so it brings down
execution time from 14 seconds to less than 6.