Make text expansion growth symmetrical
Alternate way to calculate growth of text expansion. With the old logic, if a source became
twice in length in the target, it would produce a 1.00 growth (i.e. +100%). But if it was
half the length it'd produce a -0.5 growth (i.e. -50%). This is asymmetrical because the
upper bound is +infinity and the lower bound is -1.
With the new logic, the positive value for growth is the same. However if translation has
shrunk, rather than using the shrinking ratio we use the negative of the expansion of the
target towards the source (hence producing a -infinity lower bound).