New snap command
Remove implicit source mirror and replace it with an explicit command to dump
normalized sources.

Create associated FsSnapStore to write sources to the filesystem as json files.
The modified timestamp is removed to be more diff-friendly (git doesn't carry
timestamp of files so every new clone has new dates, so committing those is
not interesting).

Fix CLI not catching exceptions.