Refactor analyzers
Made column names and groups dynamic (part of the response instead of static members) so that they can be
determined at runtime (e.g. to provide a dynamic list of languages in each column).

Simplified declaration so that driver is implied by what method is used in the analyzer
(processSegment for source and processTU for tm).

Added ExportTranslationGrid analyzer to demo getAggregateAnalysis(). Also FindInTarget.