Implement seq property to shorten guid
Config can now specify an optional seqMap filename to store a mapping between all historically used guids in the
source and assign them a sequence number. This number is then made available in the source and TU.

As a side-effect, when building the TM cache, now the latest seq, prj and notes properties are automatically updated.

Visicode: embed seq number is available.

TOS: include seq as a dashboard label if available.