Optimize source cache and snaps (phase 1)
Remove source cache and assumption that all resources will fit in memory.
Instead of mm.source.getResources() have a getResourceStats() and getResource(rs).
Add a new getAllResources() generator and in source adapters so that batching can be implemented.
Warning: TM no longer refreshes tu with latest source (if found). not sure what the impact is, if any.