Snap v2
Removed SnapSource. Now the SnapStore automatically acts as a source, if configured. It basically acts as a caching
layer. Actual sources in content types will only be used for taking the snapshots. All other operations will read from
the snapshot. During translation, if the resource filter supports resource generation from the normalized format, it
will use the snapshot as well. If not, it will read the raw resource from the source and translate that.
Snap store now stores files in chucks, organized by project. The snap command allows to specify the maximum size
of each chunk (in number of segments).