Refactor VS Code extension for v3
However, this is a dead-end. Electron is intentionally ABI-incompatible with native modules so it's impossible to have the same module (better-sqlite3) serve both CLI Node.js and VS Code.
We could switch to a WASM version of SQLite but that's not mature yet and missing features. Or wait for built-in sqlite support in Node v22 to make its way into VS Code.
Better yet, abandon VS Code (we don't get much out of it anyway) and build a standalone web-based UI ourselves.