Refactor providers for more granular troubleshooting and debugging
Throw if response cannot be parsed.
If tu throws, discard just that tu and not the whole job.
If a chunk fails, discard the chunk and not the whole job.
Remove (temporarily?) support for asynch providers in chunked provider.
Streamlined chunked provider SPI so that subclasses don't have to deal with ops.
Both raw and parsed chunks are emitted in ops now. (for troubleshooting)
Some tweaks to AI models.