fix(helpers-googlecloud): Ensure prefix is always treated as a directory (#45)
* fix(helpers-googlecloud): Ensure prefix is always treated as a directory 

it is reasonable to attempt to name gcs store named that has common prefix .e.g.,
```
 - tm-store
 - tm-store-fancy
 - tm-store-cool
```

The current list logic will cause all files across all stores to be listed for `tm-store`. A simple fix is to add `/` at the end when listing files. This also matches the string replace logic below.

* address feedback & add test

---------

Co-authored-by: Majid Valipour <mvalipour@hopper.com>