fix(helpers-lqaboss): Move lqaboss ingestion to tm syncdown
Rather than relying on a pending job waiting for the completed job, cancel the job as soon as the lqaboss file is created. Ingestion of completed jobs happens by exposing completed jobs as a tm store and syncing that down. The benefit is that completed jobs can be changed later on, and they can be replicated into different environments (which wouldn't have had the pending job required to pull them in).