feat: Improve LQA Boss
Return provider and ts in tus fetched by guid.
Support passing a jobName when starting a job.
Support specifying a quality model in LQABossProvider.
Use jobName to name job file in LQABossProvider.
Rely on providerData to pass down desired provider quality in LQABossProvider.
Increase Express body size to 10MB.
Reduce amount of data on the wire when creating a job.